Use CMD to launch bessd
We need to preserve the task affinity, hence the need for shell to be
pid 1, else BESS resets the affinity to pick the highest core from the
subset of affinitized cores.

Also we need to explicitly set the worker id to core mapping, else BESS
crashes.

https://github.com/NetSys/bess/issues/889

TODO: dynamically determine the core to map worker 0

Signed-off-by: Saikrishna Edupuganti <saikrishna.edupuganti@intel.com>